Update module 3 examples.

+++ section3/src/main/java/info/vladimirov/pluralsight/mockito/module3/ConversionRateServiceImpl.java
@@ -1,12 +1,11 @@
 package info.vladimirov.pluralsight.mockito.module3;
 
+import info.vladimirov.pluralsight.mockito.module3.api.ConversionRateService;
 import info.vladimirov.pluralsight.mockito.module3.model.AccountCurrency;
 
-import java.math.BigDecimal;
-
-public class ConversionRateServiceImpl implements ConversionRateService{
+public class ConversionRateServiceImpl implements ConversionRateService {
     @Override
-    public BigDecimal getConvertedRate(AccountCurrency currency, BigDecimal amoung) {
+    public Double getConvertedRate(AccountCurrency currency, Double amoung) {
         return null;
     }
 }
